Provide synchronous delivery-db cache utilities (#797)
* Add `_async` suffix to async deliverydb cache functions/packages

* Add synchronous database client (again)

* Add synchronous deliverydb cache module

---------
