+++ deliverydb/cache.py
@@ -1,32 +1,26 @@
-import asyncio
 import collections.abc
 import dataclasses
 import datetime
-import http
 import logging
 import traceback
 
-import aiohttp.web
-import dacite
 import sqlalchemy.exc
-import sqlalchemy.ext.asyncio as sqlasync
+import sqlalchemy.orm.session
 
-import consts
 import deliverydb.model as dm
 import deliverydb_cache.model as dcm
 import deliverydb_cache.util as dcu
-import features
 import util
 
 
 logger = logging.getLogger(__name__)
 
 
-async def update_cache_entry(
-    db_session: sqlasync.session.AsyncSession,
+def update_cache_entry(
+    db_session: sqlalchemy.orm.session.Session,
     cache_entry: dm.DBCache,
 ) -> bool:
-    if not (existing_cache_entry := await db_session.get(dm.DBCache, cache_entry.id)):
+    if not (existing_cache_entry := db_session.get(dm.DBCache, cache_entry.id)):
         # cache entry does not exist yet, hence we cannot _update_ it
         return False
 
@@ -39,33 +33,33 @@ async def update_cache_entry(
         existing_cache_entry.size = cache_entry.size
         existing_cache_entry.value = cache_entry.value
 
-        await db_session.commit()
+        db_session.commit()
         return True
     except Exception:
         stacktrace = traceback.format_exc()
         logger.error(stacktrace)
 
-        await db_session.rollback()
+        db_session.rollback()
 
     return False
 
 
-async def add_or_update_cache_entry(
-    db_session: sqlasync.session.AsyncSession,
+def add_or_update_cache_entry(
+    db_session: sqlalchemy.orm.session.Session,
     cache_entry: dm.DBCache,
 ) -> bool:
     try:
         db_session.add(cache_entry)
-        await db_session.commit()
+        db_session.commit()
         return True
 
     except sqlalchemy.exc.IntegrityError:
-        await db_session.rollback()
+        db_session.rollback()
 
         # try to update cache entry instead because it may have already existed and hence raised a
         # duplicated key error (this is the expected case if a cache entry got stale or it is
         # calculated multiple times concurrently)
-        if await update_cache_entry(
+        if update_cache_entry(
             db_session=db_session,
             cache_entry=cache_entry,
         ):
@@ -78,16 +72,16 @@ async def add_or_update_cache_entry(
         stacktrace = traceback.format_exc()
         logger.error(stacktrace)
 
-        await db_session.rollback()
+        db_session.rollback()
 
     return False
 
 
-async def find_cached_value(
-    db_session: sqlasync.session.AsyncSession,
+def find_cached_value(
+    db_session: sqlalchemy.orm.session.Session,
     id: str,
 ) -> bytes | None:
-    if not (cache_entry := await db_session.get(dm.DBCache, id)):
+    if not (cache_entry := db_session.get(dm.DBCache, id)):
         return None
 
     now = datetime.datetime.now(datetime.timezone.utc)
@@ -105,12 +99,12 @@ async def find_cached_value(
     try:
         cache_entry.last_read = now
         cache_entry.read_count = cache_entry.read_count + 1
-        await db_session.commit()
+        db_session.commit()
     except Exception:
         stacktrace = traceback.format_exc()
         logger.error(stacktrace)
 
-        await db_session.rollback()
+        db_session.rollback()
 
     return value
 
@@ -131,7 +125,7 @@ def dbcached_function(
         )
 
     def decorator(func):
-        async def wrapper(*args, **kwargs):
+        def wrapper(*args, **kwargs):
             function_name = f'{func.__module__}.{func.__qualname__}'
 
             cachable_args = tuple(
@@ -144,7 +138,7 @@ def dbcached_function(
             # remove `db_session` from kwargs to allow proper serialisation
             if not (db_session := cachable_kwargs.pop('db_session', None)):
                 logger.warning(f'Could not parse `db_session` parameter from {function_name=}')
-                return await func(*args, **kwargs)
+                return func(*args, **kwargs)
 
             shortcut_cache = cachable_kwargs.pop('shortcut_cache', False)
 
@@ -156,7 +150,7 @@ def dbcached_function(
             )
 
             if not shortcut_cache and (
-                value := await find_cached_value(
+                value := find_cached_value(
                     db_session=db_session,
                     id=descriptor.id,
                 )
@@ -167,7 +161,7 @@ def dbcached_function(
                 )
 
             start = datetime.datetime.now()
-            result = await func(*args, **kwargs)
+            result = func(*args, **kwargs)
             duration = datetime.datetime.now() - start
 
             if result in skip_values:
@@ -194,7 +188,7 @@ def dbcached_function(
                 value=value,
             )
 
-            await add_or_update_cache_entry(
+            add_or_update_cache_entry(
                 db_session=db_session,
                 cache_entry=cache_entry,
             )
@@ -206,119 +200,13 @@ def dbcached_function(
     return decorator
 
 
-def parse_shortcut_cache(
-    request: aiohttp.web.Request,
-) -> bool:
-    """
-    Parses the information, if existing cache entries should be ignored, from the given request
-    object. If the optional query parameter `shortcutCache` is set to a truthy value, it evaluates
-    to `True`. Otherwise, the value of the `Shortcut-Cache` http header is considered.
-    """
-    if util.param_as_bool(request.rel_url.query, 'shortcutCache', default=False):
-        return True
-
-    if util.param_as_bool(request.headers, 'Shortcut-Cache', default=False):
-        return True
-
-    return False
-
-
-def dbcached_route(
-    encoding_format: dcm.EncodingFormat | str = dcm.EncodingFormat.PICKLE,
-    ttl_seconds: int = 0,
-    keep_at_least_seconds: int = 0,
-    max_size_octets: int = 0,
-    skip_http_status: collections.abc.Sequence[int] = tuple(),
-):
-    if not encoding_format.startswith('pickle'):
-        raise ValueError(
-            f'Unsupported encoding format for HTTP route cache (must be pickle): {encoding_format}',
-        )
-
-    if ttl_seconds and ttl_seconds < keep_at_least_seconds:
-        raise ValueError(
-            'If time-to-live (`ttl_seconds`) and `keep_at_least_seconds` are both specified, '
-            '`ttl_seconds` must be greater or equal than `keep_at_least_seconds`.',
-        )
-
-    def decorator(func):
-        async def wrapper(*args, **kwargs):
-            # first non-keyword arg of http route functions is always the request object
-            request: aiohttp.web.Request = args[0].request
-
-            if not (db_session := request.get(consts.REQUEST_DB_SESSION)):
-                return await func(*args, **kwargs)
-
-            body = await request.json() if request.has_body else None
-
-            descriptor = dcm.CachedHTTPRoute(
-                encoding_format=encoding_format,
-                route=request.path,
-                params=dcu.normalise_and_serialise_object(request.url.query),
-                body=dcu.normalise_and_serialise_object(body) if body else None,
-            )
-
-            shortcut_cache = parse_shortcut_cache(request)
-
-            if not shortcut_cache and (
-                value := await find_cached_value(
-                    db_session=db_session,
-                    id=descriptor.id,
-                )
-            ):
-                return dcu.deserialise_cache_value(
-                    value=value,
-                    encoding_format=encoding_format,
-                )
-
-            start = datetime.datetime.now()
-            result: aiohttp.web.Response = await func(*args, **kwargs)
-            duration = datetime.datetime.now() - start
-
-            if result.status >= 400 or result.status in skip_http_status:
-                # don't cache error responses -> those might only be temporarily
-                return result
-
-            value = dcu.serialise_cache_value(
-                value=result,
-                encoding_format=encoding_format,
-            )
-
-            if max_size_octets > 0 and len(value) > max_size_octets:
-                # don't store result in cache if it exceeds max size for an individual cache entry
-                return result
-
-            now = datetime.datetime.now(datetime.timezone.utc)
-            cache_entry = dm.DBCache(
-                id=descriptor.id,
-                descriptor=util.dict_serialisation(dataclasses.asdict(descriptor)),
-                delete_after=now + datetime.timedelta(seconds=ttl_seconds) if ttl_seconds else None,
-                keep_until=now + datetime.timedelta(seconds=keep_at_least_seconds),
-                costs=int(duration.total_seconds() * 1000),
-                size=len(value),
-                value=value,
-            )
-
-            await add_or_update_cache_entry(
-                db_session=db_session,
-                cache_entry=cache_entry,
-            )
-
-            return result
-
-        wrapper.__doc__ = func.__doc__
-        return wrapper
-
-    return decorator
-
-
-async def mark_for_deletion(
-    db_session: sqlasync.session.AsyncSession,
+def mark_for_deletion(
+    db_session: sqlalchemy.orm.session.Session,
     id: str,
     delete_after: datetime.datetime | None = None,
     defer_db_commit: bool = False,
 ) -> bool:
-    if not (cache_entry := await db_session.get(dm.DBCache, id)):
+    if not (cache_entry := db_session.get(dm.DBCache, id)):
         return True
 
     if not delete_after:
@@ -328,37 +216,21 @@ async def mark_for_deletion(
         cache_entry.delete_after = delete_after
 
         if not defer_db_commit:
-            await db_session.commit()
+            db_session.commit()
         return True
     except Exception:
         stacktrace = traceback.format_exc()
         logger.error(stacktrace)
 
-        await db_session.rollback()
+        db_session.rollback()
 
     return False
 
 
-async def mark_for_deletion_task(
-    db_session: sqlasync.session.AsyncSession,
-    id: str,
-    delete_after: datetime.datetime | None = None,
-    defer_db_commit: bool = False,
-):
-    await mark_for_deletion(
-        db_session=db_session,
-        id=id,
-        delete_after=delete_after,
-        defer_db_commit=defer_db_commit,
-    )
-
-    await db_session.close()
-
-
-async def mark_function_cache_for_deletion(
+def mark_function_cache_for_deletion(
     encoding_format: dcm.EncodingFormat | str,
     function: collections.abc.Callable | str,
-    db_session: sqlasync.session.AsyncSession,
+    db_session: sqlalchemy.orm.session.Session,
     delete_after: datetime.datetime | None = None,
     defer_db_commit: bool = False,
     *args,
@@ -376,96 +248,9 @@ async def mark_function_cache_for_deletion(
         kwargs=dcu.normalise_and_serialise_object(kwargs),
     )
 
-    await mark_for_deletion(
+    mark_for_deletion(
         db_session=db_session,
         id=descriptor.id,
         delete_after=delete_after,
         defer_db_commit=defer_db_commit,
     )
-
-
-class DeliveryDBCache(aiohttp.web.View):
-    required_features = (features.FeatureDeliveryDB,)
-
-    async def delete(self):
-        """
-        ---
-        description: Mark the delivery-db cache entry with the given id for deletion.
-        tags:
-        - Artefact metadata
-        parameters:
-        - in: query
-          name: id
-          required: false
-          schema:
-            type: string
-          description:
-            The descriptor id of the cache entry which should be marked for deletion. See
-            `deliverydb_cache.model.CacheDescriptorBase` for available descriptor types and how
-            their id is composed. If not specified, the full descriptor must be passed in the body.
-        - in: query
-          name: deleteAfter
-          required: false
-          schema:
-            type: string
-          description:
-            Optional, timezone-aware iso-formated datetime to schedule the deletion of the
-            referenced cache entry. If not set, the entry will be marked for immediate deletion.
-        requestBody:
-          required: false
-          content:
-            application/json:
-              schema:
-                type: object
-          description:
-            If the descriptor id is not passed as param, this descriptor is used to calculate the
-            required id. The passed descriptor must be serialisable by one of the dataclasses which
-            inherits `deliverydb_cache.model.CacheDescriptorBase` base class.
-        responses:
-          "204":
-            description: Successful operation.
-        """
-        db_session_low_prio = self.request[consts.REQUEST_DB_SESSION_LOW_PRIO]
-        params = self.request.rel_url.query
-
-        now = datetime.datetime.now(tz=datetime.timezone.utc)
-
-        id = util.param(params, 'id', required=False)
-        delete_after_str = util.param(params, 'deleteAfter', default=now.isoformat())
-        delete_after = datetime.datetime.fromisoformat(delete_after_str)
-
-        if not id:
-            if not self.request.has_body:
-                raise aiohttp.web.HTTPBadRequest(
-                    text='Either `id` param or descriptor body must be specified.',
-                )
-
-            type_to_class = {
-                dcm.CacheValueType.COMPONENT_DESCRIPTOR: dcm.CachedComponentDescriptor,
-                dcm.CacheValueType.PYTHON_FUNCTION: dcm.CachedPythonFunction,
-                dcm.CacheValueType.HTTP_ROUTE: dcm.CachedHTTPRoute,
-            }
-
-            body = await self.request.json()
-            cache_value_type = util.get_enum_value_or_raise(body.get('type'), dcm.CacheValueType)
-
-            descriptor = dacite.from_dict(
-                data_class=type_to_class[cache_value_type],
-                data=body,
-                config=dacite.Config(
-                    cast=[dcm.CacheValueType],
-                ),
-            )
-            id = descriptor.id
-
-        asyncio.create_task(
-            mark_for_deletion_task(
-                db_session=db_session_low_prio,
-                id=id,
-                delete_after=delete_after,
-            ),
-        )
-
-        return aiohttp.web.Response(
-            status=http.HTTPStatus.NO_CONTENT,
-        )
